some fixes

+++ src/utils/file.hpp
@@ -262,7 +262,7 @@ class OutputFile {
  private:
   void FlushBuffer();
   void FlushBufferInternal();
-  void FlushBufferInternal(unsigned long to_write);
+  void FlushBufferInternal(size_t to_write);
 
   size_t SeekFile(Position position, ssize_t offset);
 
@@ -356,7 +356,7 @@ class NonConcurrentOutputFile {
  private:
   void FlushBuffer();
   void FlushBufferInternal();
-  void FlushBufferInternal(unsigned long to_write);
+  void FlushBufferInternal(size_t to_write);
 
   size_t SeekFile(Position position, ssize_t offset);
